refactor: 체력 및 데미지 값을 float에서 int로 변경

+++ Assets/02. Scipts/InGame/Monster/MonsterStats.cs
@@ -6,14 +6,14 @@ public class MonsterStats : MonoBehaviour
     [Header("몬스터 데이터")]
     [SerializeField] private MonsterData _baseData;
 
-    private float _currentHealth;
-    private float _currentMaxHealth;
-    private float _currentAttackPower;
+    private int _currentHealth;
+    private int _currentMaxHealth;
+    private int _currentAttackPower;
     private float _currentMoveSpeed;
 
-    public float CurrentHealth => _currentHealth;
-    public float MaxHealth => _currentMaxHealth;
-    public float AttackPower => _currentAttackPower;
+    public int CurrentHealth => _currentHealth;
+    public int MaxHealth => _currentMaxHealth;
+    public int AttackPower => _currentAttackPower;
     public float MoveSpeed => _currentMoveSpeed;
 
     private Coroutine _dotDamageCoroutine;
@@ -37,7 +37,7 @@ public class MonsterStats : MonoBehaviour
         _currentMoveSpeed = _baseData.MoveSpeed;
     }
 
-    public void TakeDamage(float damage)
+    public void TakeDamage(int damage)
     {
         if (_currentHealth <= 0) return;
         _currentHealth -= damage;
@@ -47,7 +47,7 @@ public class MonsterStats : MonoBehaviour
             Death();
         }
     }
-    public void TakeDotDamage(float damage, float duration, float interval)
+    public void TakeDotDamage(int damage, float duration, float interval)
     {
         if (_currentHealth <= 0) return;
         if (_dotDamageCoroutine != null)
@@ -57,7 +57,7 @@ public class MonsterStats : MonoBehaviour
         _dotDamageCoroutine = StartCoroutine(ApplyDotDamage(damage, duration, interval));
     }
 
-    private IEnumerator ApplyDotDamage(float damage, float duration, float interval)
+    private IEnumerator ApplyDotDamage(int damage, float duration, float interval)
     {
         float elapsedTime = 0.0f;
         while (elapsedTime < duration)
